Bugfix in TTeam::TTeam (-1 bug)
Changed datatype to pdf sheet info
some more should be changed...

Renamed project

+++ Ex02/tdate.cpp
@@ -1,25 +1,22 @@
 #include <iostream>
 #include <iomanip>
 #include <ctime>
-#include "TDate.h"
-#include "TTime.h"
-#include "TScore.h"
-#include "TPlayer.h"
-#include "TTeam.h"
+#include "tdate.h"
 
 using namespace std;
+
 
 TDate::TDate()
 {
     TDate::setCurrentDate();
 }
 
-TDate::TDate(int day, int month, int year)
+TDate::TDate(short day, short month, short year)
 {
     TDate::setDate(day, month, year);
 }
 
-void TDate::setDate(int day, int month, int year)
+void TDate::setDate(short day, short month, short year)
 {
     if ( day > 0 and day <=31 and month > 0 and month <=12 and
         year > 1900 and year < 3000 )
@@ -43,17 +40,17 @@ void TDate::setCurrentDate()
     TDate::year = 1900 + ltm->tm_year;
 }
 
-int TDate::getDay()
+short TDate::getDay()
 {
     return TDate::day;
 }
 
-int TDate::getMonth()
+short TDate::getMonth()
 {
     return TDate::month;
 }
 
-int TDate::getYear()
+short TDate::getYear()
 {
     return TDate::year;
 }
