+++ Ex02/ttime.cpp
@@ -1,25 +1,22 @@
 #include <iostream>
 #include <iomanip>
 #include <ctime>
-#include "TDate.h"
-#include "TTime.h"
-#include "TScore.h"
-#include "TPlayer.h"
-#include "TTeam.h"
+#include "ttime.h"
 
 using namespace std;
+
 
 TTime::TTime()
 {
     TTime::setCurrentTime();
 }
 
-TTime::TTime(int hour, int minute)
+TTime::TTime(short hour, short minute)
 {
     TTime::setTime(hour, minute);
 }
 
-void TTime::setTime(int hour, int minute)
+void TTime::setTime(short hour, short minute)
 {
     if ( hour >= 0 and hour <=23 and minute >= 0 and minute <=59 )
     {
@@ -40,12 +37,12 @@ void TTime::setCurrentTime()
     TTime::minute = ltm->tm_min;
 }
 
-int TTime::getHour()
+short TTime::getHour()
 {
     return TTime::hour;
 }
 
-int TTime::getMinute()
+short TTime::getMinute()
 {
     return TTime::minute;
 }
